+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 <!-- There is always Unreleased section on the top. Subsections (Add, Changed, Fix, Removed) should be Add as needed. -->
 ## Unreleased
+- Use custom **equality** for a `Profiler.Label`
 
 ## 3.0.0 - 2021-01-19
 - Use Lmc prefix for Assembly